KNMI Thesis
Scripts dynamics: contains scripts for computing the dynamics analysis and plotting the dynamics figures

Scripts scenarios: includes scripts to compute some figures of the scenarios part. Does not contain the methodology of the scenarios because it is private.

Scripts others: contains scripts that I did not use for the report but they plot some useful figures

Scripts preprocessing: contains the scripts and bash documents I used to preprocess CMIP6 data. HighResMIP and EC-Earth3 bis preprocessing not included